Added support for 301s between languages.
Added rudimentary support for automating 301 redirect between
languages. So, for example, if you have a route like /example which
get translated to /en/example and /es/ejemplo, these changes set up
/es/example => /es/ejemplo and /en/ejemplo => /en/example via 301
redirects. Thus, a user can simply change the locale in any URL from
the application generated by this localization, and the rest of the
URL will be corrected to the canonical form.